requirements.txt
autopep8==1.6.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
pymongo==4.1.1
python-dotenv==0.20.0
toml==0.10.2
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
